Refactored convo_lib into convo, began documentation
