--- convo-lib/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "convo-lib"
-version = "0.1.0"
-authors = ["Spencer Imbleau <spencer@imbleau.com>", "Thomas Houtart <thomas.houtart98@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-yaml-rust = "0.4"
-serde = "1.0"
-serde_yaml = "0.8"
