--- convo-lib/src/lib.rs
@@ -1,14 +1,0 @@
-// TODO: Write documentation
-/*
-#![warn(missing_docs)]
-#![warn(rustdoc::missing_crate_level_docs)]
-#![warn(rustdoc::missing_doc_code_examples)]
-#![warn(rustdoc::invalid_codeblock_attributes)]
-#![warn(rustdoc::broken_intra_doc_links)]
-*/
-
-pub mod exporter;
-pub mod link;
-pub mod node;
-pub mod parser;
-pub mod tree;
